Extracted XML doc reader and xml element handler from docxfmt.cpp to a separate file to be used in both odt and docx parsers
